Parallelize regression test jobs (#10177)
Changes:
- Use github actions strategy: `max-parallel: 3` by default
- Combine reports of all job into a single report.json
- Send metrics to Segment and Datadog while running the jobs in parallel
- Refactor: DRY the `add_comment_results_cpu` and `add_comment_results_gpu` jobs
