+++ .github/scripts/mr_generate_summary.py
@@ -1,87 +1,42 @@
 # Collect the results of the various model test runs which are done as part of
-# The model regression CI pipeline and dump them as a single file artifact.This artifact will the then be published at the end of the tests.
+# the model regression CI pipeline and dump them as a single file artifact.
+# This artifact will the then be published at the end of the tests.
 import json
 import os
+from pathlib import Path
+from typing import Dict
+
+REPORTS_DIR = Path(os.environ['REPORTS_DIR'])
+print("REPORTS_DIR", list(REPORTS_DIR.glob("*")))
 
 SUMMARY_FILE = os.environ["SUMMARY_FILE"]
-CONFIG = os.environ["CONFIG"]
-DATASET = os.environ["DATASET_NAME"]
-TYPE = os.environ["TYPE"]
-IS_EXTERNAL = os.environ["IS_EXTERNAL"]
-CONFIG_REPOSITORY = "training-data"
-CONFIG_REPOSITORY_BRANCH = os.environ["DATASET_REPOSITORY_BRANCH"]
-DATASET_REPOSITORY_BRANCH = os.environ["DATASET_REPOSITORY_BRANCH"]
-task_mapping = {
-    "intent_report.json": "intent_classification",
-    "CRFEntityExtractor_report.json": "entity_prediction",
-    "DIETClassifier_report.json": "entity_prediction",
-    "response_selection_report.json": "response_selection",
-    "story_report.json": "story_prediction",
-}
+print("SUMMARY_FILE", SUMMARY_FILE)
 
 
-def generate_json(file, task, data):
-    global IS_EXTERNAL
-    global DATASET_REPOSITORY_BRANCH
+def combine_result(result1: Dict[str, dict], result2: Dict[str, dict]) -> Dict[str, dict]:
+    import copy
+    combined_dict = copy.deepcopy(result1)
 
-    if not DATASET in data:
-        data = {DATASET: {CONFIG: {}}, **data}
-    elif not CONFIG in data[DATASET]:
-        data[DATASET] = {CONFIG: {}, **data[DATASET]}
+    for dataset, results_for_dataset in result2.items():
+        for config, res in results_for_dataset.items():
 
-    if str(IS_EXTERNAL).lower() in ("yes", "true", "t", "1"):
-        IS_EXTERNAL = True
-        DATASET_REPOSITORY_BRANCH = os.environ["EXTERNAL_DATASET_REPOSITORY_BRANCH"]
-    else:
-        IS_EXTERNAL = False
+            if dataset not in combined_dict:
+                combined_dict[dataset] = {}
 
-    data[DATASET][CONFIG] = {
-        "external_dataset_repository": IS_EXTERNAL,
-        "dataset_repository_branch": DATASET_REPOSITORY_BRANCH,
-        "config_repository": CONFIG_REPOSITORY,
-        "config_repository_branch": CONFIG_REPOSITORY_BRANCH,
-        "dataset_commit": os.environ["DATASET_COMMIT"],
-        "accelerator_type": os.environ["ACCELERATOR_TYPE"],
-        "test_run_time": os.environ["TEST_RUN_TIME"],
-        "train_run_time": os.environ["TRAIN_RUN_TIME"],
-        "total_run_time": os.environ["TOTAL_RUN_TIME"],
-        "type": TYPE,
-        **data[DATASET][CONFIG],
-    }
-
-    data[DATASET][CONFIG][task] = {**read_results(file)}
-
-    return data
-
-
-def read_results(file):
-    with open(file) as json_file:
-        data = json.load(json_file)
-
-        keys = [
-            "accuracy",
-            "weighted avg",
-            "macro avg",
-            "micro avg",
-            "conversation_accuracy",
-        ]
-        result = {key: data[key] for key in keys if key in data}
-
-    return result
+            assert config not in combined_dict[dataset]
+            combined_dict[dataset][config] = res
+    return combined_dict
 
 
 if __name__ == "__main__":
     data = {}
-    if os.path.exists(SUMMARY_FILE):
-        with open(SUMMARY_FILE) as json_file:
-            data = json.load(json_file)
 
-    for dirpath, dirnames, files in os.walk(os.environ["RESULT_DIR"]):
-        for f in files:
-            if f not in task_mapping.keys():
-                continue
+    reports_paths = list(REPORTS_DIR.glob("*/report.json"))
+    print('reports_paths', reports_paths)
 
-            data = generate_json(os.path.join(dirpath, f), task_mapping[f], data)
+    for report_path in reports_paths:
+        report_dict = json.load(open(report_path))
+        data = combine_result(data, report_dict)
 
     with open(SUMMARY_FILE, "w") as f:
         json.dump(data, f, sort_keys=True, indent=2)
