+++ .github/scripts/mr_publish_results.py
@@ -1,5 +1,6 @@
 # Send model regression test results to Segment with a summary
 # of all test results.
+# Also write them into a report file.
 import analytics
 import datetime
 import json
@@ -21,9 +22,22 @@ EXTERNAL_DATASET_REPOSITORY_BRANCH = None
 CONFIG_REPOSITORY = "training-data"
 CONFIG_REPOSITORY_BRANCH = os.environ["DATASET_REPOSITORY_BRANCH"]
 
+SUMMARY_FILE = os.environ["SUMMARY_FILE"]
+CONFIG = os.environ["CONFIG"]
+DATASET = os.environ["DATASET_NAME"]
+TYPE = os.environ["TYPE"]
+
 analytics.write_key = os.environ["SEGMENT_TOKEN"]
 
 task_mapping = {
+    "intent_report.json": "intent_classification",
+    "CRFEntityExtractor_report.json": "entity_prediction",
+    "DIETClassifier_report.json": "entity_prediction",
+    "response_selection_report.json": "response_selection",
+    "story_report.json": "story_prediction",
+}
+
+task_mapping_segment = {
     "intent_report.json": "Intent Classification",
     "CRFEntityExtractor_report.json": "Entity Prediction",
     "DIETClassifier_report.json": "Entity Prediction",
@@ -178,14 +192,68 @@ def read_results(file):
 def push_results(file_name, file):
     result = read_results(file)
     result["file_name"] = file_name
-    result["task"] = task_mapping[file_name]
+    result["task"] = task_mapping_segment[file_name]
     send_to_segment(result)
+
+
+def generate_json(file, task, data):
+    global IS_EXTERNAL
+    global DATASET_REPOSITORY_BRANCH
+
+    if not DATASET in data:
+        data = {DATASET: {CONFIG: {}}, **data}
+    elif not CONFIG in data[DATASET]:
+        data[DATASET] = {CONFIG: {}, **data[DATASET]}
+
+    if str(IS_EXTERNAL).lower() in ("yes", "true", "t", "1"):
+        IS_EXTERNAL = True
+        DATASET_REPOSITORY_BRANCH = os.environ["EXTERNAL_DATASET_REPOSITORY_BRANCH"]
+    else:
+        IS_EXTERNAL = False
+
+    data[DATASET][CONFIG] = {
+        "external_dataset_repository": IS_EXTERNAL,
+        "dataset_repository_branch": DATASET_REPOSITORY_BRANCH,
+        "config_repository": CONFIG_REPOSITORY,
+        "config_repository_branch": CONFIG_REPOSITORY_BRANCH,
+        "dataset_commit": os.environ["DATASET_COMMIT"],
+        "accelerator_type": os.environ["ACCELERATOR_TYPE"],
+        "test_run_time": os.environ["TEST_RUN_TIME"],
+        "train_run_time": os.environ["TRAIN_RUN_TIME"],
+        "total_run_time": os.environ["TOTAL_RUN_TIME"],
+        "type": TYPE,
+        **data[DATASET][CONFIG],
+    }
+
+    data[DATASET][CONFIG][task] = {**read_results(file)}
+
+    return data
+
+
+def send_all_results_to_segment():
+    for dirpath, dirnames, files in os.walk(os.environ["RESULT_DIR"]):
+        for f in files:
+            if any(f.endswith(valid_name) for valid_name in task_mapping_segment.keys()):
+                push_results(f, os.path.join(dirpath, f))
+    analytics.flush()
+
+
+def create_report_file():
+    assert not os.path.exists(SUMMARY_FILE)  # Debug
+
+    data = {}
+    for dirpath, dirnames, files in os.walk(os.environ["RESULT_DIR"]):
+        for f in files:
+            if f not in task_mapping.keys():
+                continue
+
+            data = generate_json(os.path.join(dirpath, f), task_mapping[f], data)
+
+    with open(SUMMARY_FILE, "w") as f:
+        json.dump(data, f, sort_keys=True, indent=2)
 
 
 if __name__ == "__main__":
     send_to_datadog(None)
-    for dirpath, dirnames, files in os.walk(os.environ["RESULT_DIR"]):
-        for f in files:
-            if any(f.endswith(valid_name) for valid_name in task_mapping.keys()):
-                push_results(f, os.path.join(dirpath, f))
-    analytics.flush()
+    send_all_results_to_segment()
+    create_report_file()
